rpi_verify/requirements.txt
opencv-python
pyzbar
web3
python-dotenv
